+++ src/backend/app/models/expense.py
@@ -1,6 +1,6 @@
 import datetime as dt
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer
+from sqlalchemy import DECIMAL, Column, DateTime, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
 from backend.app.core.db import Base
@@ -12,7 +12,7 @@ class Expense(Base):
     date = Column(DateTime, default=dt.datetime.now, index=True)
     category_id = Column(Integer, ForeignKey("category.id"), nullable=False)
     category = relationship("Category", back_populates="expenses", lazy="joined")
-    amount = Column(Float, nullable=False)
+    amount = Column(DECIMAL(15, 3), nullable=False)
     currency_id = Column(Integer, ForeignKey("currency.id"), nullable=True)
     currency = relationship("Currency", back_populates="expenses", lazy="joined")
