+++ src/backend/app/models/setting.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer
+from sqlalchemy import DECIMAL, Column, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
 from backend.app.core.db import Base
@@ -8,7 +8,7 @@ class Setting(Base):
     """System settings."""
 
     # ToDo: only positive
-    budget = Column(Integer, nullable=False)
+    budget = Column(DECIMAL(15, 3), nullable=False)
     default_currency_id = Column(Integer, ForeignKey("currency.id"), nullable=True)
     default_currency = relationship(
         "Currency", back_populates="default_currency", lazy="joined"
