+++ src/backend/app/schemas/setting.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, ConfigDict, Field
 
 from .currency import CurrencyDB
@@ -7,7 +9,7 @@ class SettingDB(BaseModel):
     model_config = ConfigDict(from_attributes=True)
 
     id: int
-    budget: int
+    budget: Decimal
     default_currency: CurrencyDB
 
 
@@ -16,4 +18,4 @@ class DefaultCurrency(BaseModel):
 
 
 class Budget(BaseModel):
-    budget: int = Field(..., ge=0)
+    budget: Decimal = Field(..., ge=0)
